Move API to skyplane import
